Offload AlarmReceiver's work to an IntentService.
